Sanketh_and_Garv/DashboardAnalyticsForJJ/requirements.txt
anthropic==0.89.0
langgraph==1.1.6
openpyxl==3.1.5
pandas==3.0.2
plotly==6.6.0
pytest==9.0.2
python-dotenv==1.2.2
reflex==0.8.28.post1
reportlab==4.4.10
